Refactor assistant chat into smaller modules
